Refactor App component: extract logic into useNewsAggregator hook for improved readability and maintainability
